backend/pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "eipr-backend"
version = "0.1.0"
description = "API backend for Eastern Iowa Pet Resort management platform"
authors = [{ name = "EIPR Team", email = "team@eipr.local" }]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "fastapi>=0.112.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic-settings>=2.2.1",
    "sqlalchemy>=2.0.29",
    "asyncpg>=0.29.0",
    "psycopg[binary]>=3.1.18",
    "alembic>=1.13.1",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "email-validator>=2.1.1",
    "python-dotenv>=1.0.1",
    "fastapi-pagination>=0.12.34",
    "python-multipart>=0.0.9",
    "greenlet>=3.0",
    "jinja2>=3.1.4",
    "Pillow>=10.0.0",
    "PyMySQL>=1.1.0",
    "PyYAML>=6.0",
    "cryptography>=42.0.0",
    "secure>=0.3.0",
    "fastapi-limiter>=0.1.6",
    "redis>=5.0.0",
    "asgi-correlation-id>=4.3.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.2",
    "pytest-asyncio>=0.23.7",
    "httpx>=0.27.0",
    "ruff>=0.5.0",
    "black>=24.4.0",
    "mypy>=1.8.0",
    "types-passlib>=1.7.7.4",
    "types-python-jose>=3.3.4.20240106",
    "aiosqlite>=0.20.0",
    "pre-commit>=3.7.0",
    "pytest-cov>=5.0.0"
]

[tool.setuptools.packages.find]
where = ["app"]
include = ["*"]
